Delete subtractdicts because it is complex and unused
The substractdict test is breaking on Python 3. Since (as far as I can
tell) this function is not used anywhere, I think it's better to delete
it rather than try to fix it.

(I think it was trying to pretend that keys that were bytes were the
same as keys which were strings [which is weird])